requirements.txt
# ACT XII - Purchase Predictor Requirements
Flask==3.0.0
dash==2.14.2
plotly==5.18.0
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
joblib==1.3.2
psycopg2-binary==2.9.9
python-dotenv==1.0.0
